+++ lib/chart/gantt.ex
@@ -1,5 +1,19 @@
 defmodule Contex.GanttChart do
+@moduledoc """
+Generates a Gantt Chart.
 
+Bars are drawn for each task covering the start and end time for each task. In addition, tasks can be grouped
+into categories which have a different coloured background - this is useful for showing projects that are
+in major phases.
+
+The time interval columns must be of a date time type (either `NaiveDateTime` or `DateTime`)
+
+Labels can optionally be drawn for each task (use `show_task_labels/2`) and a description for each task, including
+the time interval is generated and added as a '&lt;title&gt;' element attached to the bar. Most browsers provide
+a tooltip functionality to display the title when the mouse hovers over the containing element.
+
+By default, the first four columns of the supplied dataset are used for the category, task, start time and end time.
+"""
 alias __MODULE__
 alias Contex.{Scale, OrdinalScale, TimeScale, CategoryColourScale}
 alias Contex.Dataset
@@ -8,11 +22,23 @@ alias Contex.Utils
 
 defstruct [:dataset, :width, :height, :category_col, :task_col, :show_task_labels, :interval_cols, :time_scale, :task_scale, :padding, :category_scale, :phx_event_handler, id_col: ""]
 
+  @doc """
+  Create a new Gantt Chart definition and apply defaults.
+  """
+  @spec new(Contex.Dataset.t()) :: Contex.GanttChart.t()
   def new(%Dataset{} = dataset) do
     %GanttChart{dataset: dataset, width: 100, height: 100}
     |> defaults()
   end
 
+  @doc """
+  Sets defaults for the Gantt Chart.
+
+  The first four columns in the dataset are used for category, task, start date/time and end date/time.
+
+  Task labels are enabled by default.
+  """
+  @spec defaults(Contex.GanttChart.t()) :: Contex.GanttChart.t()
   def defaults(%GanttChart{dataset: dataset} = plot) do
     cat_col_index = 0
     task_col_index = 1
@@ -24,11 +50,15 @@ defstruct [:dataset, :width, :height, :category_col, :task_col, :show_task_label
     |> set_task_interval_cols({Dataset.column_name(dataset, start_col_index), Dataset.column_name(dataset, end_col_index)})
   end
 
-  # TODO: replace with option setting
+  @doc """
+  Show or hide labels on the bar for each task
+  """
+  @spec show_task_labels(Contex.GanttChart.t(), boolean()) :: Contex.GanttChart.t()
   def show_task_labels(%GanttChart{} = plot, show_task_labels) do
     %{plot | show_task_labels: show_task_labels}
   end
 
+  @doc false
   def set_size(%GanttChart{} = plot, width, height) do
     # We pretend to set columns to force a recalculation of scales - may be expensive.
     # We only really need to set the range, not recalculate the domain
@@ -37,6 +67,11 @@ defstruct [:dataset, :width, :height, :category_col, :task_col, :show_task_label
     |> set_task_interval_cols(plot.interval_cols)
   end
 
+  @doc """
+  Specify the columns used for category and task
+  """
+  @spec set_category_task_cols(Contex.GanttChart.t(), Contex.Dataset.column_name(), Contex.Dataset.column_name()) ::
+          Contex.GanttChart.t()
   def set_category_task_cols(%GanttChart{dataset: dataset, height: height, padding: padding} = plot, cat_col_name, task_col_name) do
     tasks = Dataset.unique_values(dataset, task_col_name)
     categories = Dataset.unique_values(dataset, cat_col_name)
@@ -50,6 +85,11 @@ defstruct [:dataset, :width, :height, :category_col, :task_col, :show_task_label
     %{plot | category_col: cat_col_name, task_col: task_col_name , task_scale: task_scale, category_scale: cat_scale}
   end
 
+  @doc """
+  Specify the columns used for start and end time of each task.
+  """
+  @spec set_task_interval_cols(Contex.GanttChart.t(), {Contex.Dataset.column_name(), Contex.Dataset.column_name()}) ::
+          Contex.GanttChart.t()
   def set_task_interval_cols(%GanttChart{dataset: dataset, width: width} = plot, {start_col, end_col}) do
     {min, _} = Dataset.column_extents(dataset, start_col)
     {_, max} = Dataset.column_extents(dataset, end_col)
@@ -61,6 +101,11 @@ defstruct [:dataset, :width, :height, :category_col, :task_col, :show_task_label
     %{plot | interval_cols: {start_col, end_col}, time_scale: time_scale}
   end
 
+  @doc """
+  Optionally specify a LiveView event handler. This attaches a `phx-click` attribute to each bar element. Note that it may
+  not work with some browsers (e.g. Safari on iOS).
+  """
+  @spec event_handler(Contex.GanttChart.t(), String.t()) :: Contex.GanttChart.t()
   def event_handler(%GanttChart{}=plot, event_handler) do
     %{plot | phx_event_handler: event_handler}
   end
@@ -69,10 +114,12 @@ defstruct [:dataset, :width, :height, :category_col, :task_col, :show_task_label
   If id_col is set it is used as the value sent by the phx_event_handler.
   Otherwise, the category and task is used
   """
+  @spec set_id_col(Contex.GanttChart.t(), Contex.Dataset.column_name()) :: Contex.GanttChart.t()
   def set_id_col(%GanttChart{}=plot, id_col_name) do
     %{plot | id_col: id_col_name}
   end
 
+  @doc false
   def to_svg(%GanttChart{time_scale: time_scale} = plot, _options) do
     time_axis = Axis.new_bottom_axis(time_scale) |> Axis.set_offset(plot.height)
     toptime_axis = Axis.new_top_axis(time_scale) |> Axis.set_offset(plot.height)
@@ -192,7 +239,5 @@ defstruct [:dataset, :width, :height, :category_col, :task_col, :show_task_label
   defp midpoint({a, b}), do: (a + b) / 2.0
   defp width({a, b}), do: abs(a - b)
   defp max_band({a1, b1}, {a2, b2}), do: {Utils.safe_min(a1, a2), Utils.safe_max(b1, b2)}
-
-
 
 end